update and share toolbar
