+++ Stanford360/Hydration/View/HydrationView.swift
@@ -1,0 +1,51 @@
+//
+// This source file is part of the Stanford 360 based on the Stanford Spezi Template Application project
+//
+// SPDX-FileCopyrightText: 2025 Stanford University
+//
+// SPDX-License-Identifier: MIT
+//
+
+@_spi(TestingSupport) import SpeziAccount
+import SwiftUI
+
+struct DailyHydrationData: Identifiable {
+	let id = UUID()
+	let dayName: String
+	let intakeOz: Double
+}
+
+struct HydrationView: View {
+	@Binding private var presentingAccount: Bool
+	
+	// MARK: - Body
+	var body: some View {
+		NavigationView {
+			ScrollView {
+				VStack(spacing: 20) {
+					HydrationTimeFrameView()
+					HydrationControlPanel()
+				}
+			}
+			.toolbar {
+				Toolbar(presentingAccount: $presentingAccount, title: "My Hydration 💧")
+			}
+			.contentShape(Rectangle())
+			.onTapGesture {
+				UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+			}
+		}
+	}
+	
+	init(presentingAccount: Binding<Bool>) {
+		self._presentingAccount = presentingAccount
+	}
+}
+
+// MARK: - Preview
+#Preview {
+	@Previewable @State var presentingAccount = false
+	
+	HydrationView(presentingAccount: $presentingAccount)
+		.environment(Stanford360Standard())
+}
