+++ Stanford360/Resources/Localizable.xcstrings
@@ -496,18 +496,6 @@
     "Move. Drink. Eat. Can you score 360 today? 🎯🏆" : {
 
     },
-    "My Active Journey 🏃‍♂️" : {
-
-    },
-    "My Dashboard" : {
-
-    },
-    "My Hydration 💧" : {
-
-    },
-    "My Protein 🍴" : {
-
-    },
     "Navigate to task details" : {
 
     },
